chore: rework integration tests to pass more ports/options to container and remove jest env overrides
seems simple to just use beforeAll/afterAll to set up the client since we know what to expect. the withContainer package didn't allow multiple ports, so modified to do so so we can eventually use 1 image for multiple files if we want or use this for other purposes. in the end, didn't really need it since we only need api or flight, technically as it is, but it isn't super complex and lets us modify it as desired