feat: change API to be singleton/instance based
Prior to this change the methods on the clients use the names
deterministic and signed to differentiate between different kinds of
instances of a model. These names expected the user to understand how
Ceramic implements account relations (i.e. using determinism).

Instead now the API uses the names singleton and instance as they
communicate the behavior of the different kinds of instances. Tests that
demonstrate the difference have been added.

Additionally the handle packages have been removed along with their
tests. Those packages where from a time when the design was to build an
aggregator in the sdk which is no longer the case. Removing them
simplifies the API the SDK provides.