feat: adds API for updating models and updating instances to use them
With this change it is now possible to update a model definition in
backwards compatible ways. Additionally when creating or updating model
instance documents a modelVersion header value can be supplied to
indicate that the instance uses the updated version of the model.
Instances will always use the init event of a model (i.e. its first
version) unless explicitly changed to a new version using the
modelVersion header.